backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
python-jose==3.3.0
passlib==1.7.4
pydantic-settings==2.6.1
python-multipart==0.0.12
email-validator==2.2.0
httpx==0.28.1
